Menu with programmes and subjects
