+++ src/containers/Masthead/MastheadContainer.jsx
@@ -30,6 +30,10 @@ import {
 import { getLocaleUrls } from '../../util/localeHelpers';
 import ErrorBoundary from '../ErrorPage/ErrorBoundary';
 import { mapMastheadData } from './mastheadHelpers';
+import {
+  getCategorizedSubjects,
+  getProgrammes,
+} from '../../util/programmesSubjectsHelper';
 
 const MastheadContainer = ({
   infoContent,
@@ -150,6 +154,8 @@ const MastheadContainer = ({
               resource={resource}
               topicResourcesByType={topicResourcesByType || []}
               locale={locale}
+              programmes={getProgrammes(locale)}
+              subjectCategories={getCategorizedSubjects(locale)}
             />
           )}
           {!hideBreadcrumb && (
