+++ src/shapes.js
@@ -271,3 +271,18 @@ export const ConceptLicenseShape = PropTypes.shape({
   src: PropTypes.string.isRequired,
   title: PropTypes.string,
 });
+
+export const SubjectCategoryShape = PropTypes.shape({
+  name: PropTypes.string.isRequired,
+  subjects: PropTypes.arrayOf(
+    PropTypes.shape({
+      name: PropTypes.string.isRequired,
+      url: PropTypes.string.isRequired,
+    }),
+  ),
+});
+
+export const ProgrammeShape = PropTypes.shape({
+  label: PropTypes.string.isRequired,
+  url: PropTypes.string.isRequired,
+});
